+++ src-tests/Process/ProcessSetCreatorTest.php
@@ -126,6 +126,19 @@ class ProcessSetCreatorTest extends \PHPUnit_Framework_TestCase
         $this->creator->createFromFiles($files, [], []);
     }
 
+    public function testShouldThrowExceptionIfAddingClassWithNameMismatchingTheFileName()
+    {
+        $files = $this->findDummyTests('WrongClassTest.php', 'InvalidTests');
+
+        $fileName = key(iterator_to_array($files->getIterator()));
+        $this->setExpectedException(
+            \RuntimeException::class,
+            'Error loading class "Lmc\Steward\Process\Fixtures\InvalidTests\ReallyWrongClassTest" from file "'
+            . $fileName . '". Make sure the class name and namespace matches the file path.'
+        );
+        $this->creator->createFromFiles($files, [], []);
+    }
+
     public function testShouldOnlyAddTestsOfGivenGroups()
     {
         $processSet = $this->creator->createFromFiles($this->findDummyTests(), ['bar', 'foo'], []);