+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Changed
 - Throw explanatory exception when attempting to load test file without any test class defined (instead of confusing `ReflectionException`).
+- Throw explanatory exception when test class cannot be instantiated (if class name/namespace doesn't match file path).
 
 ## 1.3.0 - 2016-02-26
 ### Added